fix: audioRegions are now allowed (reloading a project with a user audio would not work.
