--- src/components/video-editor/audio/sourceAudioTracks.ts
@@ -1,5 +1,0 @@
-import { getSourceTrackIdFromPath } from "@/lib/exporter/sourceTrackRoutingPolicy";
-
-export const SOURCE_AUDIO_FALLBACK_TOAST_ID = "source-audio-fallback-error";
-export const SOURCE_AUDIO_NORMALIZE_GAIN = 1.35;
-export { getSourceTrackIdFromPath };
